Add rotate-logs.sh
Very basic size-based log rotation. Can run it as a cron job. Tweak the
variables at the top of the script to control log destination and size.

Nobody should use this, but I was away from internet and had some free
time to work on another cron job that I wanted logging for, so I shaved
the yak and rolled my own. It was a fun afternoon in any event.